script for library discovery.
This uses the system LD, which is on the one hand moderately terrifying,
and yet also apparently is How Things Are Done.

Heavy on the comments.
That regexp was harrowing to write.
I'm sure it's not pleasant to try to understand at a glance, either.

Extracted this from some other scripts I've written previously
that had it directly coupled to file relocations.  Gonna try to
bring that functionality along next, but as a distinct phase of script.